🏷️ refacto(rewardvault): Update IRewardVault
- Update RewardVault's interface to respect latest RewardVault's implementation
- Extend RewardVault from IRewardVault to make sure the interface will stay accurate in the future
- Changed visibility of `rewardTokens` to internal and added `getRewardTokens` function for better access.

+++ packages/strategies/src/RewardReceiver.sol
@@ -58,7 +58,8 @@ contract RewardReceiver {
                 IERC20(token).safeIncreaseAllowance(address(rewardVault()), balance);
 
                 // Deposit rewards to the vault
-                IRewardVault(rewardVault()).depositRewards(token, balance);
+                // TODO: unsafe: cast the balance to uint128 safely
+                IRewardVault(rewardVault()).depositRewards(token, uint128(balance));
             }
         }
     }
@@ -80,7 +81,8 @@ contract RewardReceiver {
             token.safeIncreaseAllowance(address(rewardVault()), amount);
 
             // Deposit rewards to the vault
-            rewardVault().depositRewards(address(token), amount);
+            // TODO: unsafe: cast the balance to uint128 safely
+            rewardVault().depositRewards(address(token), uint128(amount));
         }
     }
 }
